% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem_ele.R
\name{get_dem_ele}
\alias{get_dem_ele}
\title{Extract Elevation from DEM}
\usage{
get_dem_ele(Data, dem_file)
}
\arguments{
\item{Data}{A data frame containing longitude and latitude coordinates.}

\item{dem_file}{Path to the DEM raster file.}
}
\value{
A data frame with an added 'elevation' column containing the extracted elevation values.
}
\description{
This function extracts elevation values from a Digital Elevation Model (DEM)
for given longitude and latitude coordinates.
}
\examples{
\dontrun{

Data <- data.frame(
   species = sample(paste0("spp_", 1:10), 500, replace = TRUE),
   year = sample(1950:2020, 500, replace = TRUE),
   month = sample(1:12, 500, replace = TRUE),
   Lon = runif(500, -10, 20),
   Lat = runif(500, 30, 70)
)

 dem_file <- "path/to/your/dem.tif"

 Data_with_elevation <- get_dem_ele(Data, dem_file)

 print(Data_with_elevation)
}

}
