% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem_ele.R
\name{get_dem_elevation}
\alias{get_dem_elevation}
\title{Extract Elevation from DEM}
\usage{
get_dem_elevation(df, dem_file)
}
\arguments{
\item{df}{A data frame containing longitude and latitude coordinates.}

\item{dem_file}{Path to the DEM raster file.}
}
\value{
A data frame with an added 'elevation' column containing the extracted elevation values.
}
\description{
This function extracts elevation values from a Digital Elevation Model (DEM)
for given longitude and latitude coordinates.
}
\examples{
# Example data frame
Data <- data.frame(
  species = sample(paste0("spp_", 1:10), 10, replace = TRUE),
  Lon = runif(10, -10, 20),
  Lat = runif(10, 30, 70)
)

# Replace 'path/to/your/dem.tif' with the actual path to your DEM file
# dem_file <- "path/to/your/dem.tif"
# Data_with_elevation <- get_dem_elevation(Data, dem_file)
# print(Data_with_elevation)

}
