% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Extract elevation from DEM}
\usage{
get_elevation(data, dem_file)
}
\arguments{
\item{data}{A \verb{data frame} containing location coordinates. Must include columns specified by \code{lat} and \code{lon}.}

\item{dem_file}{Full \code{character} path to the downloaded DEM raster file.}
}
\value{
The input data frame \code{data} with a new column  (\code{ele}) containing the extracted elevation values.
}
\description{
This function extracts elevation values from a Digital Elevation Model (DEM) based on location (\code{lon/lat}).
}
