% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_tm.R
\name{get_era5_tme}
\alias{get_era5_tme}
\title{Extract ERA5 data from NetCDF file}
\usage{
get_era5_tme(data, nc_file)
}
\arguments{
\item{data}{A data frame containing points of interest. Must include 'lon', 'lat', 'year', and 'month' columns.}

\item{nc_file}{Path to the NetCDF raster file containing the climate data.}
}
\value{
The input data frame \code{data} with a new column named (\code{tempetarure}), containing the climate values.
}
\description{
This function extracts climate variable values from an ERA5 NetCDF file (or similar structure) based on location (lon/lat) and time (year/month) for a set of data points.
}
