% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_strategy.R
\name{spp_strategy}
\alias{spp_strategy}
\title{Creates a summary table of significant and spatial/thermal trends.}
\usage{
spp_strategy(spp_trends, bonferroni = 0.05)
}
\arguments{
\item{spp_trends}{A data frame containing trend indicators per species.
It is expected to contain columns such as:
- "species": Species names.
- "Trend": Trend values.
- "t": t-statistic values.
- "pvalue": p-values.
- "Response": Response variable (e.g., Lat, Tmax).
- "Dif_t": Difference in t-statistic values.
- "Dif_pvalue": Difference in p-values.
- "n": Sample size.}

\item{bonferroni}{The Bonferroni adjusted significance threshold value.}
}
\value{
A data frame summarizing significant trends and spatial/thermal
classification of species.
}
\description{
This function takes a table of trend indicators 'spp_trends' and transforms it
into a summary table that includes information about significance, trends,
differences, and spatial/thermal classification of species.
}
\examples{
# spp_trend_result <- spp_trend(Data, spp, predictor, responses, n_min = 50)
# bonferroni = 0.05 # Recommended (0.05/length(spp))
# strategies <- spp_strategy(spp_trend_result, bonferroni)
}
