% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_trend.R
\name{spp_trend}
\alias{spp_trend}
\title{Indivudual Trend Analysis}
\usage{
spp_trend(Data, spp, predictor, responses, n_min = 50)
}
\arguments{
\item{Data}{A data frame containing the variables for the model.}

\item{spp}{A vector of species names.}

\item{predictor}{A vector of predictor variable names.}

\item{responses}{In this case date is response.}

\item{n_min}{Value correspondign to minumun presences of a specie.}
}
\value{
A data frame with trend statistics, including:
\itemize{
\item \code{Response}: The name of the variable analyzed.
\item \code{Trend}: The slope of the linear model.
\item \code{t}: The t-statistic of the model.
\item \code{pvalue}: The p-value of the trend.
\item \code{ci_95_max}: Upper bound of the 95\% confidence interval.
\item \code{ci_95_min}: Lower bound of the 95\% confidence interval.
}
}
\description{
This function fits a linear model to analyze individual trends over time and comparing with general data trend
}
\examples{

Data <- data.frame(
  species = sample(paste0("spp_", 1:10), 500, replace = TRUE),
  year = sample(1900:2024, 500, replace = TRUE),
  month = sample(1:12, 500, replace = TRUE),
  Lon = runif(500, -10, 20),
  Lat = runif(500, 30, 70),
  Tmx = rnorm(500, 15, 10),
  Tmn = rnorm(500, 10, 8))

Data$year_month  = Data$year + Data$month * 0.075

predictor <- "year_month" # predictor <- "year" in case month is not available
responses <- c("Lat", "Lon", "TMmx", "TMmn")
spp <- unique(Data$species)

general_trend_result <- spp_trend(Data, spp, predictor, responses, n_min = 50)

}
