% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_trend_world.R
\name{spp_trend_world}
\alias{spp_trend_world}
\title{This function fits a linear model to analyze individual trends over time,
comparing with general data trend, and includes longitude transformation
and hemisphere detection.}
\usage{
spp_trend_world(Data, spp, predictor, responses, n_min = 50)
}
\arguments{
\item{Data}{A data frame containing the variables for the model, including
'species', 'year', 'month', 'Lon', 'Lat', 'Tmx', 'Tmn'.}

\item{spp}{A vector of species names.}

\item{predictor}{A vector of predictor variable names (e.g., "year_month").}

\item{responses}{A vector of response variable names (e.g., "Lat", "Lon", "Tmx").}

\item{n_min}{Minimum number of presences required for a species.}
}
\value{
A data frame with trend statistics, including:
\itemize{
\item \code{species}: Species name.
\item \code{response}: The name of the variable analyzed.
\item \code{trend}: The slope of the linear model.
\item \code{t}: The t-statistic of the model.
\item \code{pvalue}: The p-value of the trend.
\item \code{ci_95_max}: Upper bound of the 95\% confidence interval.
\item \code{ci_95_min}: Lower bound of the 95\% confidence interval.
\item \code{dif_t}: Difference in t-statistic values.
\item \code{dif_pvalue}: Difference in p-values.
\item \code{hemisphere}: Detected hemisphere ("N", "S", "Both" or "Unknown").
}
}
\description{
This function fits a linear model to analyze individual trends over time,
comparing with general data trend, and includes longitude transformation
and hemisphere detection.
}
\examples{
Data <- data.frame(
  species = sample(paste0("spp_", 1:10), 500, replace = TRUE),
  year = sample(1900:2024, 500, replace = TRUE),
  month = sample(1:12, 500, replace = TRUE),
  Lon = runif(500, -10, 20),
  Lat = runif(500, 30, 70),
  Tmx = rnorm(500, 15, 10),
  Tmn = rnorm(500, 10, 8)
)
Data$year_month <- Data$year + Data$month * 0.075
predictor <- "year_month"
responses <- c("Lat", "Lon", "Tmx", "Tmn")
spp <- unique(Data$species)
general_trend_result <- spp_trend_world(Data, spp, predictor, responses, n_min = 50)

}
