% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SppTrend_example.R
\name{SppTrend_example}
\alias{SppTrend_example}
\title{Open the SppTrend example}
\usage{
SppTrend_example()
}
\value{
Opens the PDF in the default system viewer.
}
\description{
This function locates and opens the PDF example included in the package.
}
