% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SppTrend_guide.R
\name{SppTrend_guide}
\alias{SppTrend_guide}
\title{Open the SppTrend user guide}
\usage{
SppTrend_guide()
}
\value{
Opens the PDF in the default system viewer.
}
\description{
This function locates and opens the PDF user guide included in the package.
}
