% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elevation.R
\name{extract_elevation}
\alias{extract_elevation}
\title{Extract elevation from DEM}
\usage{
extract_elevation(data, dem_file)
}
\arguments{
\item{data}{A data frame containing location coordinates. Must include columns specified by 'lon_col' and 'lat_col'.}

\item{dem_file}{Full path to the downloaded DEM raster file.}
}
\value{
The input data frame \code{data} with a new column  (\code{ele}) containing the extracted elevation values.
}
\description{
This function extracts elevation values from a Digital Elevation Model (DEM) based on location (\code{lon/lat}).
}
