% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Extract elevation from DEM}
\usage{
get_elevation(data, dem_file)
}
\arguments{
\item{data}{A \verb{data frame} containing species records. Must include \code{lon}, \code{lat}, \code{year}, and \code{month} columns.}

\item{dem_file}{Full \code{character} path to the downloaded DEM raster file.}
}
\value{
The input data frame \code{data} with a new column  (\code{ele}) containing the extracted elevation values.
}
\description{
This function retrieves elevation values from a Digital Elevation Model (DEM) based on their geographic coordinates (\code{lon/lat}).
}
