% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_tme.R
\name{get_era5_tme}
\alias{get_era5_tme}
\title{Indivudual Trend Analysis}
\usage{
get_era5_tme(Data, nc_file, month_col = NULL)
}
\arguments{
\item{Data}{A data frame containing the variables for the model.}

\item{responses}{In this case date is response.}

\item{predictor}{A vector of predictor variable names.}

\item{spp}{A vector of species names.}

\item{n_min}{Value correspondign to minumun presences of a specie.}
}
\value{
Data data frame with Mean Temperature in K
}
\description{
This function fits a linear model to analyze individual trends over time and comparing with general data trend
}
\examples{
\dontrun{

Data <- data.frame(
   species = sample(paste0("spp_", 1:10), 500, replace = TRUE),
   year = sample(1950:2020, 500, replace = TRUE),
   month = sample(1:12, 500, replace = TRUE),
   Lon = runif(500, -10, 20),
   Lat = runif(500, 30, 70)
)

nc_file <- "C:/A_TRABAJO/A_JORGE/SPP_VIRTUALES/MARIPOSAS/TEST_Spp_Trends/era5_1940_2023.nc"

Data_with_Tme <- get_era5_tme(Data, nc_file, month_col = "month")

print(Data_with_Tme)
}
}
