% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_tme.R
\name{get_era5_tme}
\alias{get_era5_tme}
\title{Indivudual Trend Analysis}
\usage{
get_era5_tme(df, nc_file, month_col = NULL)
}
\arguments{
\item{Data}{A data frame containing the variables for the model.}

\item{responses}{In this case date is response.}

\item{predictor}{A vector of predictor variable names.}

\item{spp}{A vector of species names.}

\item{n_min}{Value correspondign to minumun presences of a specie.}
}
\value{
A data frame with trend statistics.
}
\description{
This function fits a linear model to analyze individual trends over time and comparing with general data trend
}
\examples{

Data <- data.frame(
species = sample(paste0("spp_", 1:10), 10, replace = TRUE),
year = sample(1950:2023, 10, replace = TRUE),
month = sample(1:12, 10, replace = TRUE),
Lon = runif(10, -10, 20),
Lat = runif(10, 30, 70)
)

Data$year_month  = Data$month * 0.075
Data$year_month  = Data$year + Data$year_month

Data_with_Tme <- get_era5_tme(Data, nc_file, month_col = "month")

}
