% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_tm.R
\name{get_era5_tme}
\alias{get_era5_tme}
\title{Extract ERA5 data from NetCDF file}
\usage{
get_era5_tme(data, nc_file)
}
\arguments{
\item{data}{A data frame containing points of interest. Must include \code{lon}, \code{lat}, \code{year}, and \code{month} columns.}

\item{nc_file}{Path to the downloaded ERA5-Land raster (\code{.nc}) file.}
}
\value{
The input data frame \code{data} with a new column named (\code{tme}), containing the temperature values.
}
\description{
This function extracts climate variable values from an ERA5 NetCDF file based on location (\code{lon/lat}) and time (\code{year/month}).
}
