% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_strategy.R
\name{spp_strategy}
\alias{spp_strategy}
\title{Creates a summary table of significant and spatial/thermal trends.}
\usage{
spp_strategy(spp_trends, bonferroni = 0.05)
}
\arguments{
\item{spp_trends}{A data frame containing trend indicators per species.
It is expected to contain columns such as:
- "species": Species names.
- "Trend": Trend values.
- "t": t-statistic values.
- "pvalue": p-values.
- "Response": Response variable (e.g., Lat, Tmax).
- "Dif_t": Difference in t-statistic values.
- "Dif_pvalue": Difference in p-values.
- "n": Sample size.}

\item{bonferroni}{The Bonferroni adjusted significance threshold value.}
}
\value{
A data frame summarizing significant trends and spatial/thermal
classification of species, including:
\itemize{
\item \code{Response}: The name of the variable analyzed.
\item \code{species}: Species names.
\item \code{Trend}: Trend values.
\item \code{t}: t-statistic values.
\item \code{pvalue}: p-values.
\item \code{Response}: Response variable (e.g., Lat, Tmx).
\item \code{Dif_t}: Difference in t-statistic values.
\item \code{Dif_pvalue}: Difference in p-values.
\item \code{n}: Sample size.
}
}
\description{
This function takes a table of trend indicators 'spp_trends' and transforms it
into a summary table that includes information about significance, trends,
differences, and spatial/thermal classification of species.
}
\examples{

spp_trends_results <- data.frame(
  species = paste0("spp_", 1:10),
  response = rep("Lat", 10),
  trend = runif(10, -0.5, 0.5),
  t = runif(10, -2, 2),
  pvalue = runif(10, 0, 1),
  ci_95_max = runif(10, 0.05, 0.2),
  ci_95_min = runif(10, -0.1, 0.05),
  dif_t = runif(10, -1, 1.5),
  dif_pvalue = runif(10, 0.001, 0.9),
  n = runif(10, 40, 60))
spp <- unique(spp_trends_results$species)
bonferroni = 0.05/length(spp)
spp_strategy_results <- spp_strategy(spp_trends_results, bonferroni = 0.05)

}
