% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_trend.R
\name{spp_trend}
\alias{spp_trend}
\title{spp_trend}
\usage{
spp_trend(Data, spp, predictor, responses, n_min = 50)
}
\arguments{
\item{Data}{A data frame containing the variables for the model, including
'species', 'year', 'month', 'Lon', 'Lat', 'Tmx', 'Tmn'.}

\item{spp}{A vector of species names.}

\item{predictor}{A vector of predictor variable names (e.g., "year_month").}

\item{responses}{A vector of response variable names (e.g., "Lat", "Lon",
"Tmx").}

\item{n_min}{Minimum number of presences required for a species in each
hemisphere (or globally for species in both hemispheres).}
}
\value{
A data frame with trend statistics, including:
\itemize{
\item \code{species}: Species name.
\item \code{responses}: The name of the variable analyzed.
\item \code{trend}: The slope of the linear model.
\item \code{t}: The t-statistic of the model.
\item \code{pvalue}: The p-value of the trend.
\item \code{ci_95_max}: Upper bound of the 95\% confidence interval.
\item \code{ci_95_min}: Lower bound of the 95\% confidence interval.
\item \code{dif_t}: Difference in t-statistic values (comparison of species trend to general trend).
\item \code{dif_pvalue}: Difference in p-values (comparison of species trend to general trend).
\item \code{n}: Number of datapoints used for the specific species/hemisphere trend.
\item \code{hemisphere}: Detected hemisphere ("North", "South", or "Both" for global comparison).
}
}
\description{
This function fits a linear model to analyze individual trends
over time, comparing with general data trend, and includes longitude
transformation and hemisphere detection. For species distributed across
both hemispheres, it also compares their overall trend with the global data trend.
}
\details{
The function analyzes trends for each species in the provided data,
considering different response variables and a specified predictor. It
accounts for the antimeridian issue when analyzing longitude trends and
separates the analysis by hemisphere. It also compares the trend of each
species with a general trend calculated from the entire hemisphere's data.
For species found in both the Northern and Southern Hemispheres, the function
additionally compares the trend of the species (using all occurrences) with
the trend of the entire dataset (all hemispheres combined).
}
\examples{
\dontrun{
Data <- data.frame(
   species = sample(paste0("spp_", 1:10), 500, replace = TRUE),
   year = sample(1950:2020, 500, replace = TRUE),
   month = sample(1:12, 500, replace = TRUE),
   Lon = runif(500, -10, 20),
   Lat = runif(500, 30, 70),
   Tmx = rnorm(500, 15, 10),
   Tmn = rnorm(500, 10, 8)
)

Data$year_month <- Data$year + Data$month * 0.075

predictor <- "year_month"
responses <- c("Lat", "Lon", "Tmx", "Tmn")

spp <- unique(Data$species)

spp_trend_result <- spp_trend(Data, spp, predictor, responses, n_min = 50)

print(spp_trend_result)
}

}
