% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_trend.R
\name{spp_trend}
\alias{spp_trend}
\title{Indivudual Trend Analysis}
\usage{
spp_trend(Data, spp, predictor, responses, n_min = 50)
}
\arguments{
\item{Data}{A data frame containing the variables for the model.}

\item{spp}{A vector of species names.}

\item{predictor}{A vector of predictor variable names.}

\item{responses}{In this case date is response.}

\item{n_min}{Value correspondign to minumun presences of a specie.}
}
\value{
A data frame with trend statistics.
}
\description{
This function fits a linear model to analyze individual trends over time and comparing with general data trend
}
\examples{

Data <- data.frame(
  species = sample(paste0("spp_", 1:10), 500, replace = TRUE),
  year = sample(1900:2024, 500, replace = TRUE),
  month = sample(1:12, 500, replace = TRUE),
  Lon = runif(500, -10, 20),
  Lat = runif(500, 30, 70),
  Tmx = rnorm(500, 15, 10),
  Tmn = rnorm(500, 10, 8))

Data$year_month  = Data$month * 0.075
Data$year_month  = Data$year + Data$year_month

predictor <- "year_month"
responses <- c("Lat", "Lon", "TMmx", "TMmn")
spp <- unique(Data$species)

general_trend_result <- spp_trend(Data, spp, predictor, responses, n_min = 50)

}
