% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SppTrend-package.R
\docType{package}
\name{SppTrend-package}
\alias{SppTrend-package}
\alias{SppTrend}
\title{SppTrend: Analyzing linear trends in species occurrence data}
\description{
Provides functions to analyze trends in time series of species abundance or occurrence data, particularly focusing on the impact of environmental factors like temperature.

The \code{SppTrend} package provides a methodology to analyze how species occurrences
change over time, particularly in relation to geographic and environmental factors.
It facilitates the development of explanatory hypotheses about the impact of
environmental shifts on species assemblages by analyzing historical presence
data that includes temporal and geographic information.
}
\details{
\subsection{Methodology}{

\code{SppTrend} assumes that observed species occurrences reflect a temporal sequence of changes in response to environmental drivers.

The analysis uses:
\itemize{
\item \strong{Predictors}: Sampling date (e.g., year or year-month decimals).
\item \strong{Responses}: Geographic location (latitude and longitude) and environmental factors (elevation and temperature).
}
}

\subsection{Workflow}{

\code{SppTrend} provides a structured workflow for analyzing these trends:
\enumerate{
\item \strong{Environmental data integration (optional)}: Enhance occurrence records
with environmental context using functions like \code{\link{get_era5_tme}} (temperature)
or \code{\link{extract_elevation}} (elevation).
\item \strong{Overall trend estimation}: Calculate the average temporal trend of
selected response variables across the entire dataset using \code{\link{overall_trend}}.
This establishes a baseline for comparison.
\item \strong{Individual trend analysis}: Determine specific temporal trends for each
species and response variable using \code{\link{spp_trend}}. This compares
individual species' responses to the overall trend via interaction models.
\item \strong{Ecological strategy classification}: Categorize species into distinct
ecological strategies based on the significance and direction of their
trends relative to the baseline using \code{\link{spp_strategy}}.
}
}

\subsection{More details}{

Project home and source code: \url{https://github.com/MarioMingarro/SppTrend}
}
}
\author{
\strong{Maintainer}: Mario Mingarro \email{mario_mingarro@mncn.csic.es} (\href{https://orcid.org/0000-0003-3977-7944}{ORCID})

Authors:
\itemize{
  \item Jorge M. Lobo (\href{https://orcid.org/0000-0002-3152-4769}{ORCID})
  \item Emilio García-Roselló (\href{https://orcid.org/0000-0003-0305-770X}{ORCID})
}

}
\keyword{internal}
