% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SppTrend_example.R
\name{SppTrend_example}
\alias{SppTrend_example}
\title{Open the SppTrend example}
\usage{
SppTrend_example(format = "html")
}
\arguments{
\item{format}{A \code{character} string, either "pdf" or "html" (default).}
}
\value{
Opens the PDF or html in the default system viewer.
}
\description{
This function locates and opens the PDF example included in the package.
}
